feat(frontend): move poll page to root
