+++ frontend/app/poll/page.tsx
@@ -1,155 +1,5 @@
-'use client';
-import { supabase } from '@/utils/supabaseClient';
-import { useEffect, useState } from 'react';
-import type { Session } from '@supabase/supabase-js';
-
-interface Game {
-  id: number;
-  name: string;
-  count: number;
-  nicknames: string[];
-}
-
-interface Poll {
-  id: number;
-  games: Game[];
-}
+import { redirect } from 'next/navigation';
 
 export default function PollPage() {
-  const [poll, setPoll] = useState<Poll | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [session, setSession] = useState<Session | null>(null);
-  const [selected, setSelected] = useState<number | null>(null);
-  const [submitting, setSubmitting] = useState(false);
-
-  const fetchPoll = async () => {
-    setLoading(true);
-    const { data: pollData, error: pollError } = await supabase
-      .from('polls')
-      .select('*')
-      .order('created_at', { ascending: false })
-      .limit(1)
-      .maybeSingle();
-
-    if (pollError || !pollData) {
-      setLoading(false);
-      return;
-    }
-
-    const { data: games } = await supabase.from('games').select('id, name');
-    const { data: votes } = await supabase
-      .from('votes')
-      .select('game_id, user_id')
-      .eq('poll_id', pollData.id);
-    const { data: users } = await supabase
-      .from('users')
-      .select('id, username');
-
-    const userMap =
-      users?.reduce((acc: Record<number, string>, u) => {
-        acc[u.id] = u.username;
-        return acc;
-      }, {}) || {};
-
-    const counts: Record<number, number> = {};
-    const nicknames: Record<number, string[]> = {};
-
-    votes?.forEach((v) => {
-      counts[v.game_id] = (counts[v.game_id] || 0) + 1;
-      if (!nicknames[v.game_id]) nicknames[v.game_id] = [];
-      const name = userMap[v.user_id];
-      if (name) nicknames[v.game_id].push(name);
-    });
-
-    const results =
-      games?.map((g) => ({
-        id: g.id,
-        name: g.name,
-        count: counts[g.id] || 0,
-        nicknames: nicknames[g.id] || [],
-      })) || [];
-
-    setPoll({ id: pollData.id, games: results });
-    setLoading(false);
-  };
-
-  useEffect(() => {
-    fetchPoll();
-    supabase.auth.getSession().then(({ data: { session } }) => {
-      setSession(session);
-    });
-    const { data: { subscription } } = supabase.auth.onAuthStateChange(
-      (_event, sess) => {
-        setSession(sess);
-      }
-    );
-    return () => subscription.unsubscribe();
-  }, []);
-
-  const handleVote = async () => {
-    if (!poll || selected === null) return;
-    setSubmitting(true);
-    const token = session?.access_token;
-
-    const username =
-      session?.user.user_metadata.preferred_username ||
-      session?.user.user_metadata.name ||
-      session?.user.user_metadata.full_name ||
-      session?.user.user_metadata.nickname ||
-      session?.user.email;
-
-    await fetch('/api/vote', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        ...(token ? { Authorization: `Bearer ${token}` } : {}),
-      },
-      body: JSON.stringify({
-        poll_id: poll.id,
-        game_id: selected,
-        username,
-      }),
-    });
-    setSelected(null);
-    await fetchPoll();
-    setSubmitting(false);
-  };
-
-  if (loading) return <div className="p-4">Loading...</div>;
-  if (!poll) return <div className="p-4">No poll available.</div>;
-
-  return (
-    <main className="p-4 max-w-xl mx-auto space-y-4">
-      <h1 className="text-2xl font-semibold">Current Poll</h1>
-      <ul className="space-y-2">
-        {poll.games.map((game) => (
-          <li key={game.id} className="border p-2 rounded space-y-1">
-            <label className="flex items-center space-x-2">
-              <input
-                type="radio"
-                name="game"
-                value={game.id}
-                checked={selected === game.id}
-                onChange={() => setSelected(game.id)}
-              />
-              <span>{game.name}</span>
-              <span className="font-mono">{game.count}</span>
-            </label>
-            <ul className="pl-4 list-disc">
-              {game.nicknames.map((name) => (
-                <li key={name}>{name}</li>
-              ))}
-            </ul>
-          </li>
-        ))}
-      </ul>
-      <button
-        className="px-4 py-2 bg-purple-600 text-white rounded disabled:opacity-50"
-        disabled={selected === null || submitting || !session}
-        onClick={handleVote}
-      >
-        {submitting ? 'Voting...' : 'Vote'}
-      </button>
-    </main>
-  );
+  redirect('/');
 }
